# src/compute-plane-services/nvsnap/internal/server/openapi.yaml
...
openapi: 3.1.0
info:
  title: NvSnap API
  summary: GPU checkpoint and restore for Kubernetes workloads.
  description: |
    The NvSnap REST API. Drives checkpoint and restore of GPU
    workloads (CUDA, PyTorch, vLLM, SGLang, TRT-LLM) across nodes
    and clusters without modifying the application.

    ## API stability tiers

    Endpoints in this document fall into three groups:

    - **Public** — `Checkpoints`, `Restores`, `Retention`, `Audit`,
      `Cluster`, `System`, `Storage`, `Observability`. Stable; semver
      applies. Safe to integrate against from external tools and CI
      pipelines.

    - **Realtime** — the WebSocket stream at `/api/v1/ws`. Stable
      transport (RFC 6455 WebSocket over HTTP/1.1) and stable envelope
      shape (`{type, data}`). Topic names and message payloads may
      evolve as new operations grow new event streams.

    - **Internal** — endpoints tagged `Internal` are consumed by the
      nvsnap-agent fleet (peer-fanout catalog routing, blob-upload
      callbacks) or by the embedded UI (demo flow). They are
      documented here for completeness and operator debugging. They
      **may change without notice** between releases. Do not
      integrate against them from external code.

    ## Out of scope

    - The `nvsnap-agent` HTTP API on each node (port 8081) — separate
      internal cluster API, not in this spec.
    - The `nvsnap-blobstore` HTTP API (port 9000) — internal CAS
...
  version: "0.1.0"
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/v1/checkpoints/{id}:
    get:
      tags: [Checkpoints]
      operationId: getCheckpoint
      summary: Fetch a single checkpoint
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
        - in: query
          name: namespace
          description: Optional — required only when the same ID exists in multiple namespaces.
          schema: { type: string }
      responses:
        '200':
          description: |
            Tries the live agent inventory first; falls back to the
            Kubernetes CR. Returns `CheckpointAgent` when the agent
            holds the dump, otherwise `CheckpointCR`.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CheckpointAgent'
                  - $ref: '#/components/schemas/CheckpointCR'
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Checkpoints]
      operationId: deleteCheckpoint
      summary: Delete a checkpoint
      description: |
        Routes to the agent that holds the dump on disk and deletes
        both the on-disk dump and the catalog row.
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
      responses:
        '204':
          description: Deleted
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/checkpoints/{id}/sources:
    get:
      tags: [Internal]
      operationId: getCheckpointSources
      summary: "Internal — peer + blob URI for cascade restore"
      description: |
        Read by `nvsnap-agent`'s restore-side cascade before pulling a
        checkpoint. Returns the prioritized peer list (most-recently
        -seen first) plus the optional blob-store URI as a tier-3
        fallback.
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckpointSources' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/checkpoints/{id}/peer-add:
    post:
      tags: [Internal]
      operationId: peerAddCheckpoint
      summary: "Internal — register a node as a peer"
      description: |
        Called by `nvsnap-agent` immediately after it lands a local copy
        of the checkpoint (capture-source or successful cascade
        fetch). Idempotent.
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeerRegisterRequest' }
      responses:
        '204':
          description: Peer registered.
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/checkpoints/{id}/peer-remove:
    post:
      tags: [Internal]
      operationId: peerRemoveCheckpoint
      summary: "Internal — unregister a node as a peer"
      description: |
        Called by `nvsnap-agent` when it evicts a checkpoint from its
        local cache. Idempotent.
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [node_name]
              properties:
                node_name: { type: string }
      responses:
        '204':
          description: Peer removed (or already absent).
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/v1/checkpoints/register:
    post:
      tags: [Internal]
      operationId: registerCheckpoint
      summary: "Internal — upsert a checkpoint row in the catalog"
      description: |
        Called by `nvsnap-agent` after a direct API-driven capture (one
        that didn't go through `nvsnap-server`'s CRD flow) so the
        catalog has a row to anchor peer-add / blob-uploaded /
        sources routing on. Idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterCheckpointRequest' }
      responses:
        '204':
          description: Catalog row upserted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/checkpoints/{id}/blob-uploaded:
    post:
      tags: [Internal]
      operationId: blobUploadedCheckpoint
      summary: "Internal — record successful blob-store upload"
      description: |
        Called by `nvsnap-agent`'s background uploader after every file
        in the dump dir is durably stored in `nvsnap-blobstore`. Sets
        the catalog's `blob_uri` so cascade restore can fall back to
        the blob store when no peer is reachable.
      parameters:
        - $ref: '#/components/parameters/CheckpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blob_uri]
              properties:
                blob_uri:
                  type: string
                  description: Base URL of the nvsnap-blobstore that received the upload.
      responses:
        '204':
          description: Blob URI recorded.
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    CheckpointId:
      in: path
      name: id
      required: true
      description: Opaque checkpoint identifier returned by `createCheckpoint`.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Server-side error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    CheckpointCR:
      type: object
      description: |
        Flattened view of a `nvsnap.io/v1alpha1 GPUCheckpoint` Kubernetes
        Custom Resource — `spec` and `status` merged into a single
        object. Returned for `source=crd` (the default) and as a
        fallback for `getCheckpoint`.
      properties:
        id:             { type: string, description: CR `metadata.name`. }
        namespace:      { type: string }
        createdAt:      { type: string, format: date-time, description: CR `metadata.creationTimestamp`. }
        # spec
        podName:        { type: string }
        containerName:  { type: string }
        leaveRunning:   { type: boolean }
        storageClass:   { type: string }
        retentionDays:  { type: integer }
        # status — written by nvsnap-server's reconciler
        phase:
          type: string
          description: |
            Pipeline phase. Values currently written by nvsnap-server:
            `InProgress`, `Completed`, `Failed`. The CRD schema also
            allows `Pending`, `Running`, `Restoring`, `Restored` but
            those aren't emitted by this build.
        message:        { type: string }
        startTime:      { type: string, format: date-time }
        completionTime: { type: string, format: date-time }
        checkpointPath: { type: string }
        checkpointSize: { type: integer, format: int64 }
        nodeName:       { type: string, description: Node the dump landed on. }
    CheckpointAgent:
      type: object
      description: |
        Per-node `metadata.json` returned by `nvsnap-agent` and aggregated
        by `nvsnap-server` for `source=agent`. The agent reads each
        checkpoint directory on disk; `nvsnap-server` adds `nodeName` and
        `source: "agent"` before returning.
      additionalProperties: true
      properties:
        # nvsnap-server-added fields
        nodeName: { type: string }
        source:   { type: string, enum: [agent] }
        # CheckpointMetadata fields (subset — see internal/agent/checkpoint.go)
        version:         { type: string, description: Metadata schema version, e.g. `v1.7`. }
        id:              { type: string }
        createdAt:       { type: string, format: date-time }
        checkpointSize:  { type: integer, format: int64 }
        durationSeconds: { type: number, format: double }
        podName:         { type: string }
        podNamespace:    { type: string }
        containerName:   { type: string }
        containerID:     { type: string }
        containerImage:  { type: string }
        containerPid:    { type: integer }
        rootfs:          { type: string }
        # Other fields (skipped, mappedFiles, openFiles, cuda, criuOptions,
        # restoreHints, compression, integrity, replayMounts, …) are
        # implementation detail — present but not contract.
...
    CheckpointSources:
      type: object
      properties:
        checkpoint_id: { type: string }
        peers:
          type: array
          items:
            type: object
            properties:
              node_name:     { type: string }
              agent_url:     { type: string }
              registered_at: { type: string, format: date-time }
              last_seen:     { type: string, format: date-time }
        blob_uri:
          type: string
          description: Empty until the agent's background uploader fires `/blob-uploaded`.
...
    PeerRegisterRequest:
      type: object
      required: [node_name, agent_url]
      properties:
        node_name: { type: string }
        agent_url: { type: string, description: Base URL of the agent on the registering node (e.g. `http://<host-ip>:8081`). }
    RegisterCheckpointRequest:
      type: object
      required: [checkpoint_id, namespace, pod_name, node_name]
      properties:
        checkpoint_id:   { type: string }
        namespace:       { type: string }
        pod_name:        { type: string }
        container_name:  { type: string }
        container_image: { type: string }
        node_name:       { type: string }
        checkpoint_path: { type: string }
        checkpoint_size: { type: integer, format: int64 }
        status:
          type: string
          default: Completed
          description: "Defaults to `Completed`. Use `Failed` to record a known-failed capture."
        has_gpu:         { type: boolean }
        duration_secs:   { type: number, format: double }
...